requirements.txt
Deprecated==1.2.12
Flask==2.0.1
GitHub-Flask==3.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
Werkzeug==2.0.1
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
google-api-core==1.30.0
google-api-python-client==2.8.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-auth==1.30.2
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
itsdangerous==2.0.1
oauthlib==3.1.1
packaging==20.9
protobuf==3.17.3
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pyparsing==2.4.7
pytz==2021.1
requests-oauthlib==1.3.0
requests==2.25.1
rsa==4.7.2
six==1.16.0
uritemplate==3.0.1
urllib3==1.26.5
wrapt==1.12.1
